fix: 修复找不到rpm包索引时删除包失败的问题 #158

+++ src/backend/rpm/biz-rpm/src/main/kotlin/com/tencent/bkrepo/rpm/artifact/repository/RpmLocalRepository.kt
@@ -657,7 +657,7 @@ class RpmLocalRepository(
                         size = (index.metadata?.get("size") as String).toLong(),
                         timestamp = index.metadata?.get("timestamp") as String,
                         openChecksum = RpmChecksum(index.metadata?.get("openChecksum") as String),
-                        openSize = (index.metadata?.get("openSize") as String).toInt()
+                        openSize = (index.metadata?.get("openSize") as String).toLong()
                     )
                 } else {
                     RepoGroup(
